requirements.txt
flask
flask-cors
python-dotenv
requests
flask-login
flask-bcrypt
flask-wtf